python/requirements-demo1.txt
torch>=2.1
transformers>=4.40
numpy
matplotlib
pyyaml
tqdm